src/requirements.txt
Django>=5.2.8
djangorestframework
django-cors-headers
django-filter
psycopg[binary]
python-dotenv
djangorestframework-simplejwt
Faker
pytest
pytest-django
ruff